+++ crates/accelerate/src/synthesis/clifford/mod.rs
@@ -41,7 +41,7 @@ fn synth_clifford_greedy(py: Python, clifford: PyReadonlyArray2<bool>) -> PyResu
         GreedyCliffordSynthesis::new(tableau.view()).map_err(QiskitError::new_err)?;
     let (num_qubits, clifford_gates) = greedy_synthesis.run().map_err(QiskitError::new_err)?;
 
-    CircuitData::from_standard_gates(py, num_qubits as u32, clifford_gates, Param::Float(0.0))
+    CircuitData::from_standard_gates(py, num_qubits, clifford_gates, Param::Float(0.0))
 }
 
 /// Generate a random Clifford tableau.
@@ -77,7 +77,7 @@ fn synth_clifford_bm(py: Python, clifford: PyReadonlyArray2<bool>) -> PyResult<C
     let tableau = clifford.as_array();
     let (num_qubits, clifford_gates) =
         synth_clifford_bm_inner(tableau).map_err(QiskitError::new_err)?;
-    CircuitData::from_standard_gates(py, num_qubits as u32, clifford_gates, Param::Float(0.0))
+    CircuitData::from_standard_gates(py, num_qubits, clifford_gates, Param::Float(0.0))
 }
 
 pub fn clifford(m: &Bound<PyModule>) -> PyResult<()> {
