+++ crates/accelerate/src/synthesis/linear/pmh.rs
@@ -191,5 +191,5 @@ pub fn synth_cnot_count_full_pmh(
             )
         });
 
-    CircuitData::from_standard_gates(py, num_qubits as u32, instructions, Param::Float(0.0))
+    CircuitData::from_standard_gates(py, num_qubits, instructions, Param::Float(0.0))
 }
