+++ crates/accelerate/src/synthesis/linear_phase/mod.rs
@@ -37,7 +37,7 @@ pub(crate) mod cz_depth_lnn;
 fn synth_cz_depth_line_mr(py: Python, mat: PyReadonlyArray2<bool>) -> PyResult<CircuitData> {
     let view = mat.as_array();
     let (num_qubits, lnn_gates) = cz_depth_lnn::synth_cz_depth_line_mr_inner(view);
-    CircuitData::from_standard_gates(py, num_qubits as u32, lnn_gates, Param::Float(0.0))
+    CircuitData::from_standard_gates(py, num_qubits, lnn_gates, Param::Float(0.0))
 }
 
 pub fn linear_phase(m: &Bound<PyModule>) -> PyResult<()> {
