+++ crates/accelerate/src/unitary_synthesis.rs
@@ -290,7 +290,7 @@ fn py_run_main_loop(
             };
         }
         if !(packed_instr.op.name() == "unitary"
-            && packed_instr.op.num_qubits() >= min_qubits as u32)
+            && packed_instr.op.num_qubits() >= min_qubits)
         {
             out_dag.push_back(py, packed_instr)?;
             continue;
